requirements.txt
cmake
beautifulsoup4
boto3
pandas
pyarrow
psycopg2-binary
python-dotenv
